Disable Anilist API.
NO I WONT JOIN YOUR STUPID DISCORD SERVER!
