+++ error.vue
@@ -7,35 +7,34 @@
         <!-- <p class="mt-7 text-7xl font-bold" v-else>{{ error.statusCode }}</p> -->
         
         <p class="mt-7 text-6xl">Oops.</p>
-        <p class="mt-7" v-if="isRouteAnime == true" >{{ animeMessage }}</p>
-        <p class="mt-7" v-else>{{ error.message }}</p>
+        <p class="mt-7">{{ error.message }}</p>
         <a href="/"><button class=" mt-7 text-sm font-semibold inline-block py-2 px-4 rounded-lg text-slate-300 uppercase last:mr-0 mr-4 border border-gradient-to-r from-cyan-500 to-blue-500">Go home!</button></a>
     </div>
 </template>
 
 <script setup>
     defineProps(['error'])
-    const route = useRoute()
-    let isRouteAnime;
-    let animeMessage;
-    route.name == 'anime' ? isRouteAnime = true : isRouteAnime = false
-    if (isRouteAnime == true) {
-        // check if Anilist is 200
-        fetch("https://anilist.co")
-        .then(res => {
-            console.log(res);
-            if (res.statusCode == 200) {
-                return animeMessage = "Something went wrong on our side."
-            } else {
-                return animeMessage = "Anilist seems to be down. Please try again  later."
-            }
+    // const route = useRoute()
+    // let isRouteAnime;
+    // let animeMessage;
+    // route.name == 'anime' ? isRouteAnime = true : isRouteAnime = false
+    // if (isRouteAnime == true) {
+    //     // check if Anilist is 200
+    //     fetch("https://anilist.co")
+    //     .then(res => {
+    //         console.log(res);
+    //         if (res.statusCode == 200) {
+    //             return animeMessage = "Something went wrong on our side."
+    //         } else {
+    //             return animeMessage = "Anilist seems to be down. Please try again  later."
+    //         }
 
-        })
-        .catch(err => {
-            console.log(err)
-            animeMessage = "Seems like Anilist is down. Please try again later."
-        })
-    }
+    //     })
+    //     .catch(err => {
+    //         console.log(err)
+    //         animeMessage = "Seems like Anilist is down. Please try again later."
+    //     })
+    // }
 </script>
 
 <style>
